chore(node): add licenses, rename file to _import
